mysite/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
mysql-connector-python==8.0.5
mysqlclient==2.1.1
protobuf==3.20.3
pycodestyle==2.10.0
sqlparse==0.4.3
tzdata==2022.7
